DOI and three new functions
This is a minor update of the package. In this version I fixed a minor bug in the valProbggplot and val.prob.ci.2 functions.

Additionally, I added the CRAN DOI to the documentation and three new functions to the packages. Two of them are infix operators to run code as a background job using RStudio/Posit. Consequently, the examples are set to donttest since they can only be executed within RStudio/Posit.